remove ChartsDemoPageComponent and move ChatComponent inside /packages/charts/src/lib/components
