+++ packages/charts/src/lib/components/chat.component/chat.component.html
