+++ packages/charts/src/lib/components/chat.component/chat.component.ts
@@ -1,9 +1,16 @@
 import { Component } from '@angular/core';
-import { DataService } from '../../common/services/data.service';
-import { ChatResponse } from '../../common/models/chat-response.model'; // لو عاملينه
+import { DataService } from '../../services/data.service';
+
+
+interface ChatResponse {
+  chartType: string;
+  title: string;
+  values: any;
+}
+
 
 @Component({
-  selector: 'app-chat',
+  selector: 'chat',
   standalone: true,
   templateUrl: './chat.component.html',
   styleUrls: ['./chat.component.css']
@@ -13,8 +20,10 @@ export class ChatComponent {
   message: string = '';
   chatHistory: string[] = [];
 
+  
+
   constructor(private dataService: DataService) {
-    // TypeScript يعرف إن history هو string[]
+
     this.dataService.chatHistory$.subscribe((history: string[]) => {
       this.chatHistory = history;
     });
@@ -27,10 +36,10 @@ export class ChatComponent {
   sendMessage(msg: string): void {
     if (!msg.trim()) return;
 
-    // أضف الرسالة للـ service
+
     this.dataService.addChatMessage(msg);
 
-    // حاول تحديث chart
+  
     try {
       const chatResponse: ChatResponse = JSON.parse(msg);
       this.dataService.updateChartData(chatResponse);
