+++ packages/charts/src/lib/components/loader/loader.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: 'app-loader',
+  selector: 'loader',
   standalone: true,
   templateUrl: './loader.component.html',
   styleUrls: ['./loader.component.css']
