+++ packages/charts/src/lib/components/toggle/toggle.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 
 @Component({
-  selector: 'app-toggle',
+  selector: 'toggle',
   standalone: true,
   templateUrl: './toggle.component.html',
   styleUrls: ['./toggle.component.css']
