+++ packages/charts/src/pages/components/chart.component/chart.component.html
@@ -29,5 +29,5 @@
   <div echarts [options]="chartOptions" class="chart-card__chart"></div>
 
   <!-- Chat Component -->
-  <app-chat (messageSent)="handleChatMessage($event)"></app-chat>
+  <chat (messageSent)="handleChatMessage($event)"></chat>
 </div>
