--- packages/charts/src/pages/components/charts-demo-page/charts-demo-page.component.css
@@ -1,12 +1,0 @@
-.charts-demo-container {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 24px;
-  padding: 16px;
-}
-
-@media (max-width: 768px) {
-  .charts-demo-container {
-    grid-template-columns: 1fr; /* stacked for mobile */
-  }
-}
