--- packages/charts/src/pages/components/charts-demo-page/charts-demo-page.component.html
@@ -1,7 +1,0 @@
-<div class="charts-demo-container">
-  <!-- Chat floating -->
-  <app-chat></app-chat>
-
-  <!-- Chart -->
-  <app-charts [chatResponse]="dataService.latestChartResponse"></app-charts>
-</div>
